added acces on route & added playlist crud

+++ src/Controller/AlbumController.php
@@ -10,6 +10,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\ExpressionLanguage\Expression;
 
 #[Route('/album')]
 final class AlbumController extends AbstractController
@@ -17,6 +18,9 @@ final class AlbumController extends AbstractController
     #[Route(name: 'all-albums', methods: ['GET'])]
     public function index(AlbumRepository $albumRepository): Response
     {
+        $this->denyAccessUnlessGranted(new Expression(
+            'is_granted("ROLE_USER")'
+        ));
         return $this->render('album/index.html.twig', [
             'albums' => $albumRepository->findAll(),
         ]);
@@ -25,6 +29,9 @@ final class AlbumController extends AbstractController
     #[Route('/new', name: 'app_album_new', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted(new Expression(
+            'is_granted("ROLE_ARTIST")'
+        ));
         $album = new Album();
         $form = $this->createForm(AlbumType::class, $album);
         $form->handleRequest($request);
@@ -45,6 +52,10 @@ final class AlbumController extends AbstractController
     #[Route('/{id}', name: 'album-show', methods: ['GET'])]
     public function show(Album $album): Response
     {
+        $this->denyAccessUnlessGranted(new Expression(
+            'is_granted("ROLE_USER")'
+        ));
+
         return $this->render('album/show.html.twig', [
             'album' => $album,
         ]);
@@ -53,6 +64,10 @@ final class AlbumController extends AbstractController
     #[Route('/{id}/edit', name: 'album-edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Album $album, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted(new Expression(
+            'is_granted("ROLE_ARTIST")'
+        ));
+
         $form = $this->createForm(AlbumType::class, $album);
         $form->handleRequest($request);
 
@@ -71,6 +86,10 @@ final class AlbumController extends AbstractController
     #[Route('/{id}', name: 'album-delete', methods: ['POST'])]
     public function delete(Request $request, Album $album, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted(new Expression(
+            'is_granted("ROLE_ARTIST")'
+        ));
+
         if ($this->isCsrfTokenValid('delete'.$album->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($album);
             $entityManager->flush();
