+++ src/Controller/ArtistController.php
@@ -10,6 +10,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\ExpressionLanguage\Expression;
 
 #[Route('/artist')]
 final class ArtistController extends AbstractController
@@ -17,6 +18,10 @@ final class ArtistController extends AbstractController
     #[Route(name: 'all-artists', methods: ['GET'])]
     public function index(ArtistRepository $artistRepository): Response
     {
+        $this->denyAccessUnlessGranted(new Expression(
+            'is_granted("ROLE_USER")'
+        ));
+
         return $this->render('artist/index.html.twig', [
             'artists' => $artistRepository->findAll(),
         ]);
@@ -25,6 +30,9 @@ final class ArtistController extends AbstractController
     #[Route('/new', name: 'artist-add', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted(new Expression(
+            'is_granted("ROLE_ADMIN")'
+        ));
         $artist = new Artist();
         $form = $this->createForm(ArtistType::class, $artist);
         $form->handleRequest($request);
@@ -45,6 +53,10 @@ final class ArtistController extends AbstractController
     #[Route('/{id}', name: 'artist-show', methods: ['GET'])]
     public function show(Artist $artist): Response
     {
+        $this->denyAccessUnlessGranted(new Expression(
+            'is_granted("ROLE_USER")'
+        ));
+
         return $this->render('artist/show.html.twig', [
             'artist' => $artist,
         ]);
@@ -53,6 +65,10 @@ final class ArtistController extends AbstractController
     #[Route('/{id}/edit', name: 'artist-edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Artist $artist, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted(new Expression(
+            'is_granted("ROLE_ADMIN")'
+        ));
+
         $form = $this->createForm(ArtistType::class, $artist);
         $form->handleRequest($request);
 
@@ -71,6 +87,10 @@ final class ArtistController extends AbstractController
     #[Route('/{id}', name: 'artist-delete', methods: ['POST'])]
     public function delete(Request $request, Artist $artist, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted(new Expression(
+            'is_granted("ROLE_ADMIN")'
+        ));
+
         if ($this->isCsrfTokenValid('delete'.$artist->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($artist);
             $entityManager->flush();
