+++ src/Controller/SongController.php
@@ -10,6 +10,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\ExpressionLanguage\Expression;
 
 #[Route('/song')]
 final class SongController extends AbstractController
@@ -17,6 +18,10 @@ final class SongController extends AbstractController
     #[Route(name: 'all-songs', methods: ['GET'])]
     public function index(SongRepository $songRepository): Response
     {
+        $this->denyAccessUnlessGranted(new Expression(
+            'is_granted("ROLE_USER")'
+        ));
+
         return $this->render('song/index.html.twig', [
             'songs' => $songRepository->findAll(),
         ]);
@@ -25,6 +30,10 @@ final class SongController extends AbstractController
     #[Route('/new', name: 'song-add', methods: ['GET', 'POST'])]
     public function new(Request $request, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted(new Expression(
+            'is_granted("ROLE_ARTIST")'
+        ));
+
         $song = new Song();
         $form = $this->createForm(SongType::class, $song);
         $form->handleRequest($request);
@@ -45,6 +54,10 @@ final class SongController extends AbstractController
     #[Route('/{id}', name: 'song-show', methods: ['GET'])]
     public function show(Song $song): Response
     {
+        $this->denyAccessUnlessGranted(new Expression(
+            'is_granted("ROLE_USER")'
+        ));
+
         return $this->render('song/show.html.twig', [
             'song' => $song,
         ]);
@@ -53,6 +66,10 @@ final class SongController extends AbstractController
     #[Route('/{id}/edit', name: 'song-edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Song $song, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted(new Expression(
+            'is_granted("ROLE_ARTIST")'
+        ));
+
         $form = $this->createForm(SongType::class, $song);
         $form->handleRequest($request);
 
@@ -71,6 +88,10 @@ final class SongController extends AbstractController
     #[Route('/{id}', name: 'song-delete', methods: ['POST'])]
     public function delete(Request $request, Song $song, EntityManagerInterface $entityManager): Response
     {
+        $this->denyAccessUnlessGranted(new Expression(
+            'is_granted("ROLE_ARTIST")'
+        ));
+
         if ($this->isCsrfTokenValid('delete'.$song->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($song);
             $entityManager->flush();
